record 정렬시킨 후 불러옴
date로 정렬시킴
