+++ Backend/src/main/java/com/pathsnap/Backend/User/Controller/UserController.java
@@ -3,6 +3,7 @@ package com.pathsnap.Backend.User.Controller;
 import com.pathsnap.Backend.User.Dto.Res.LocationResDTO;
 import com.pathsnap.Backend.User.Dto.Req.UserUpdateReqDTO;
 import com.pathsnap.Backend.User.Dto.Res.UserResDTO;
+import com.pathsnap.Backend.User.Service.LocationService;
 import com.pathsnap.Backend.User.Service.UserService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.*;
@@ -13,6 +14,10 @@ public class UserController {
 
     @Autowired
     private UserService userService;
+
+    @Autowired
+    private LocationService locationService;
+
 
     // 프로필 정보 불러오기
     @GetMapping("/{userId}")
@@ -31,6 +36,6 @@ public class UserController {
     // 여행 이미지 가져오기
     @GetMapping("/location/{userId}")
     public LocationResDTO getLocation(@PathVariable String userId) {
-        return userService.getLocations(userId);
+        return locationService.getLocations(userId);
     }
 }
